[modbus_controller] Tidy includes and narrow the payload_to_float offset
Not pushed with the reviewed work - held back for whichever pull request lands
next, or to fold in if a reviewer asks for it.

The three platform headers name std::span without including <span>, and no
longer use std::vector at all. payload_to_float() advertises a size_t offset
but hands it to a helper taking uint8_t, so it cannot carry what the type
promises; uint8_t is what it actually accepts, and the resolved offset it is
called with is uint8_t anyway.

The raw-decode lambda in the component test config was described as the
documented style. It calls a helper this release deprecates, on purpose, to pin
that the compatibility overload still takes a span - so the comment now says
that, including that the warning it raises is what a user on the old helper
sees.

+++ esphome/components/modbus_controller/modbus_controller.h
@@ -462,7 +462,7 @@ class ModbusController final : public PollingComponent, public modbus::ModbusCli
  * @param item SensorItem object
  * @return float value of data
  */
-inline float payload_to_float(std::span<const uint8_t> data, const SensorItem &item, size_t offset) {
+inline float payload_to_float(std::span<const uint8_t> data, const SensorItem &item, uint8_t offset) {
   int64_t number = modbus::helpers::payload_to_number(data, item.sensor_value_type, offset, item.bitmask).value_or(0);
 
   float float_value;
